[z0714] Use for Transport::Update arg uint32 type as for other world objects
(based on master commit 1f87b28)

(based on commit ab3e380)

+++ src/game/MapManager.cpp
@@ -246,7 +246,7 @@ MapManager::Update(uint32 diff)
     }
 
     for (TransportSet::iterator iter = m_Transports.begin(); iter != m_Transports.end(); ++iter)
-        (*iter)->Update(i_timer.GetCurrent());
+        (*iter)->Update((uint32)i_timer.GetCurrent());
 
     i_timer.SetCurrent(0);
 }
